+++ zebra-chain/src/tests/vectors.rs
@@ -1,9 +1,15 @@
 //! Network methods for fetching blockchain vectors.
 //!
 
-use std::collections::BTreeMap;
+use std::{collections::BTreeMap, ops::RangeBounds};
 
-use crate::{block::Block, parameters::Network, serialization::ZcashDeserializeInto};
+use crate::{
+    amount::Amount,
+    block::Block,
+    parameters::Network,
+    serialization::ZcashDeserializeInto,
+    transaction::{UnminedTx, VerifiedUnminedTx},
+};
 
 use zebra_test::vectors::{
     BLOCK_MAINNET_1046400_BYTES, BLOCK_MAINNET_653599_BYTES, BLOCK_MAINNET_982681_BYTES,
@@ -28,6 +34,39 @@ impl Network {
         } else {
             TESTNET_BLOCKS.iter()
         }
+    }
+
+    /// Returns iterator over verified unmined transactions in the provided block height range.
+    pub fn unmined_transactions_in_blocks(
+        &self,
+        block_height_range: impl RangeBounds<u32>,
+    ) -> impl DoubleEndedIterator<Item = VerifiedUnminedTx> {
+        let blocks = self.block_iter();
+
+        // Deserialize the blocks that are selected based on the specified `block_height_range`.
+        let selected_blocks = blocks
+            .filter(move |(&height, _)| block_height_range.contains(&height))
+            .map(|(_, block)| {
+                block
+                    .zcash_deserialize_into::<Block>()
+                    .expect("block test vector is structurally valid")
+            });
+
+        // Extract the transactions from the blocks and wrap each one as an unmined transaction.
+        // Use a fake zero miner fee and sigops, because we don't have the UTXOs to calculate
+        // the correct fee.
+        selected_blocks
+            .flat_map(|block| block.transactions)
+            .map(UnminedTx::from)
+            // Skip transactions that fail ZIP-317 mempool checks
+            .filter_map(|transaction| {
+                VerifiedUnminedTx::new(
+                    transaction,
+                    Amount::try_from(1_000_000).expect("invalid value"),
+                    0,
+                )
+                .ok()
+            })
     }
 
     /// Returns blocks indexed by height in a [`BTreeMap`].
